Optimization for qubitize
Optionnally alternate implementation order of the phasewalk operator.
This makes it a bit more obvious to the optimizer that most swaps
between layers can be cancelled, leading to slightly more efficient
circuit after transpilation.
It shaved off 5 gates of depth on an 8 qubit 2 logical layer test
circuit for ibmq_montreal.